The products component has been improved.
Fixed a bug in styles.
Improved Layout.
