+++ src/components/App/App.jsx
@@ -1,10 +1,38 @@
-import AppRoutes from '../Routes/AppRoutes';
+import { Outlet } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { fetchProducts } from '../../features/products/productsSlice';
 
+import Header from '../Header/Header';
+import Footer from '../Footer/Footer';
+import Sidebar from '../Sidebar/Sidebar';
+import Home from '../Home/Home';
+
+function Layout() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(fetchProducts());
+  }, [dispatch]);
+  return (
+    <div className="app">
+      <Header />
+      <div className="container">
+        <Sidebar />
+        <Outlet />
+      </div>
+      <Footer />
+    </div>
+  );
+}
 function App() {
   return (
-    <div>
-      <AppRoutes />
-    </div>
+    <Routes>
+      <Route path="/" element={<Layout />}>
+        <Route index element={<Home />} />
+      </Route>
+    </Routes>
   );
 }
