+++ src/components/Home/Home.jsx
@@ -1,11 +1,13 @@
+import { useSelector } from 'react-redux';
 import Poster from '../Poster/Poster';
 import Products from '../Products/Products';
 
 function Home() {
+  const products = useSelector((state) => state.products.items);
   return (
     <>
       <Poster />
-      <Products />
+      <Products title="Trending" products={products} amount={5} />
     </>
   );
 }
