--- src/components/Layout/Layout.jsx
@@ -1,28 +1,0 @@
-import { Outlet } from 'react-router-dom';
-import { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { fetchProducts } from '../../features/products/productsSlice';
-
-import Header from '../Header/Header';
-import Footer from '../Footer/Footer';
-import Sidebar from '../Sidebar/Sidebar';
-
-function Layout() {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(fetchProducts());
-  }, [dispatch]);
-  return (
-    <div className="wrapper">
-      <Header />
-      <main className="container">
-        <Sidebar />
-        <Outlet />
-      </main>
-      <Footer />
-    </div>
-  );
-}
-
-export default Layout;
