+++ src/styles/Footer.module.css
@@ -3,7 +3,6 @@
   align-items: center;
   justify-content: space-between;
   margin-top: 20px;
-  margin-bottom: 10px;
   padding: 24px;
   background: var(--dark);
   border-radius: 6px;
