+++ src/styles/Header.module.css
@@ -27,6 +27,7 @@
   width: 36px;
   height: 36px;
   border-radius: 50%;
+  background: var(--dark);
   background-repeat: no-repeat;
   background-position: center;
   background-size: cover;
@@ -86,8 +87,8 @@
 }
 
 .account svg {
-  width: 32px;
-  height: 32px;
+  width: 24px;
+  height: 24px;
 }
 
 .icon-fav {
@@ -121,7 +122,7 @@
   border-radius: 50%;
   font-weight: 700;
   font-size: 10px;
-  color: var(--bg);
+  color: var(--light);
   text-align: center;
   line-height: 16px;
 }
