+++ src/styles/Products.module.css
@@ -69,7 +69,7 @@
   font-weight: 700;
   font-size: 20px;
   line-height: 24px;
-  color: var(--violet-dark);
+  color: var(--green-dark);
 }
 
 .oldPrice {
