+++ src/styles/Sidebar.module.css
@@ -1,7 +1,7 @@
 .sidebar {
-  width: 244px;
+  width: 304px;
   padding: 24px;
-  height: 100%;
+  height: 423px;
   display: flex;
   flex-direction: column;
 }
@@ -12,8 +12,7 @@
 }
 
 .menu {
-  margin-top: 12px;
-  margin-bottom: 12px;
+  margin-top: 32px;
   display: flex;
   flex-direction: column;
   row-gap: 12px;
