+++ src/styles/index.css
@@ -1,14 +1,3 @@
+@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;700;900&family=Sofia+Sans:wght@300;500;700&display=swap");
 @import url(./variables.css);
 @import url(./global.css);
-
-.wrapper {
-    width: 1274px;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-  }
-  
-  main {
-    flex-grow: 1;
-  }
