+++ src/styles/variables.css
@@ -4,8 +4,8 @@
   --dark-sea: #576067;
   --grey: #b8b8b8;
   --white: #fff;
-  --green: #32d893;
-  --green-dark: #114625;
+  --green: #57c681;
+  --green-dark: #158b5a;
   --light: #f6f6f7;
   --sea: #616e74;
 }
